refactor: refactor folder
