+++ domain/account/ui/LoginForm/LoginForm.tsx
@@ -1,0 +1,10 @@
+const LoginForm = () => {
+  const coba = "hallo";
+  return (
+    <div>
+      <h1>ini login form {coba}</h1>
+    </div>
+  );
+};
+
+export default LoginForm;
