+++ domain/account/ui/LoginForm/index.ts
@@ -1,0 +1,3 @@
+import LoginForm from "./LoginForm";
+
+export default LoginForm;
