+++ domain/account/ui/index.ts
@@ -1,0 +1,4 @@
+import LoginForm from "./LoginForm";
+
+// eslint-disable-next-line import/prefer-default-export
+export { LoginForm };
