+++ domain/shared/styles/globals.css
