+++ domain/shared/ui/MuiIcons/MuiIcons.tsx
