+++ domain/shared/ui/MuiIcons/index.ts
