+++ domain/shared/ui/TextInput/index.ts
